+++ src/command-center.ts
@@ -1,8 +1,11 @@
+import * as path from "path";
+
 import { commands, Disposable, SourceControl, SourceControlResourceGroup, window, workspace } from "vscode";
 
 import { StatusKind } from "node-svn";
 
 import { client } from "./client";
+import { writeError, writeTrace } from "./output";
 import subscriptions from "./subscriptions";
 import { SvnResourceState } from "./svn-resource-state";
 import workspaceManager from "./workspace-manager";
@@ -32,6 +35,57 @@ class CommandCenter {
                 await window.showTextDocument(document);
             }
         }));
+
+        this.disposable.add(commands.registerCommand("svn.checkout", this.checkout));
+    }
+
+    private async checkout() {
+        const url = await window.showInputBox({
+            ignoreFocusOut: true,
+            prompt: "Checkout url",
+        });
+
+        if (url === undefined)
+            return;
+
+        let target = await window.showInputBox({
+            ignoreFocusOut: true,
+            prompt: "Checkout path",
+        });
+
+        if (target === undefined)
+            return;
+
+        while (!path.isAbsolute(target)) {
+            const folders = workspace.workspaceFolders;
+            if (folders === undefined || folders.length !== 1) {
+                const item = "Retry";
+                const selected = await window.showErrorMessage(`Can't determinate absolute path from relative path ${target}`, item);
+                switch (selected) {
+                    case item:
+                        target = await window.showInputBox({
+                            ignoreFocusOut: true,
+                            prompt: "Checkout path",
+                        });
+
+                        if (target === undefined)
+                            return;
+
+                        break;
+                    default:
+                        return;
+                }
+            } else {
+                target = path.resolve(folders[0].uri.fsPath, target);
+            }
+        }
+
+        try {
+            const revision = await client.checkout(url, target);
+            writeTrace(`checkout("${url}", "${target}")`, revision);
+        } catch (err) {
+            writeError(`checkout("${url}", "${target}")`, err);
+        }
     }
 
     private async update(e?: SourceControl) {